+++ YagomMarket/YagomMarketTests/Domain/UseCases/AddNextProductPageUseCaseTests.swift
@@ -10,33 +10,10 @@ import XCTest
 
 final class AddNextProductPageUseCaseTests: XCTestCase {
     
-    class ProductRepositoryMock: ProductsRepository {
-        var hasNext: Bool = true
-        
-        func fetchProductsList(pageNumber: Int,
-                               itemPerPage: Int,
-                               searchValue: String?) async throws -> ProductListResponseDTO {
-            return ProductListResponseDTO.toMockData(hasNext: hasNext)
-        }
-        
-        func fetchProductDetail(productId: Int) async throws -> ProductDetail {
-            return ProductDetail.toMockData()
-        }
-        
-        func fetchProductsQuery(keyword: String) async throws -> [String] {
-            return []
-        }
-        
-        func editProductDetail(with editModel: ProductEditRequestDTO,
-                               productId: Int) async throws {}
-        
-        func deleteProduct(productId: Int) async throws {}
-    }
-    
     func test_hasNext가true일때_값을성공적으로받아오는지() async throws {
         // given
         var response: [ProductCell]?
-        let productRepository = ProductRepositoryMock()
+        let productRepository = ProductsRepositoryMock()
         let useCase = DefaultAddNextProductPageUseCase(productsRepository: productRepository)
         
         // when
@@ -49,7 +26,7 @@ final class AddNextProductPageUseCaseTests: XCTestCase {
     func test_hasNext가false일때_noHasNext오류를반환하는지() async throws {
         // given
         let expectationError = ProductsRepositoryError.noNextPage
-        let productRepository = ProductRepositoryMock()
+        let productRepository = ProductsRepositoryMock()
         productRepository.hasNext = false
         let useCase = DefaultAddNextProductPageUseCase(productsRepository: productRepository)
