+++ YagomMarket/YagomMarketTests/Domain/UseCases/HandleLikedProductUseCaseTests.swift
@@ -10,41 +10,11 @@ import XCTest
 
 final class HandleLikedProductUseCaseTests: XCTestCase {
     
-    class FirestoreServiceMock<E: Entity>: FirestoreService {
-        typealias T = E
-        var firestore: [Entity] = []
-        var readCallCount = 0
-        var updateCallCount = 0
-        
-        func create<T>(collectionId: String,
-                       documentId: String,
-                       entity: T) async throws where T : YagomMarket.Entity {
-        }
-        
-        func read<T>(collectionId: String,
-                     documentId: String,
-                     entity: T) async throws -> T where T : YagomMarket.Entity {
-            readCallCount += 1
-            return UserProfile(vendorName: "wongbing",
-                               email: "test@naver.com",
-                               likedProductIds: [10,12,14]) as! T
-        }
-        
-        func update<T>(collectionId: String,
-                       documentId: String,
-                       to entity: T) async throws where T : YagomMarket.Entity {
-            updateCallCount += 1
-        }
-        
-        func delete(collectionId: String,
-                    documentId: String) async throws {}
-    }
-    
     func test_UseCase가실행될때_FirestoreService의_read메서드와_update메서드가실행되는지() async throws {
         // given
         let expectationReadCallCount = 1
         let expectationUpdateCallCount = 1
-        let firestoreService = FirestoreServiceMock<UserProfile>()
+        let firestoreService = FirestoreServiceMock()
         let useCase = HandleLikedProductUseCase(firestoreService: firestoreService)
         
         // when
