+++ YagomMarket/YagomMarketTests/Domain/UseCases/SigninUseCaseTests.swift
@@ -10,22 +10,6 @@ import FirebaseAuth
 @testable import YagomMarket
 
 final class SigninUseCaseTests: XCTestCase {
-
-    class FirebaseAuthServiceMock: FirebaseAuthService {
-        var createUserCallCount = 0
-        var signInCallCount = 0
-        
-        func createUser(email: String, password: String) async throws -> AuthDataResult? {
-            createUserCallCount += 1
-            return nil
-        }
-        
-        func signIn(email: String, password: String) async throws -> AuthDataResult? {
-            signInCallCount += 1
-            
-            return nil
-        }
-    }
     
     func test_UseCase를실행할때_FirebaseAuthService의_signIn메서드가실행되는지() async throws {
         // given
